Add pretend option for testing.

+++ send_survey.sh
@@ -39,20 +39,21 @@ BODYTEMPLATEFILE=email_body.txt
 SUBJECTTEMPLATE="community study"
 
 usage () {
-	echo
-	echo "Usage: $0 <recipientfile>"
-	echo
-	echo "BE CAREFUL! WRONG USAGE OF THIS SCRIPT CAN RESULT IN SERIOUS DAMAGE!"
-	echo
-	echo "The recipientfile contains one line for each recipient."
-	echo
-	echo "Each line of the recipientfile looks like this:"
-	echo "\"email address\" \"survey url\" \"project\" \"revision\""
-	echo
-	echo "This would be a valid line:"
-	echo "\"john.doe@example.com\" \"https://example.com/link/to/survey.php?some.params=42&other.params=42\" \"Perl\" \"v5.19.0\""
-	echo
-	echo "Also, make sure to configure the script and mutt before executing it!"
+	echo >&2
+	echo >&2 "Usage: $0 [-p] <recipientfile>"
+	echo >&2 "   -p: pretend. Do not really send mails."
+	echo >&2
+	echo >&2 "BE CAREFUL! WRONG USAGE OF THIS SCRIPT CAN RESULT IN SERIOUS DAMAGE!"
+	echo >&2
+	echo >&2 "The recipientfile contains one line for each recipient."
+	echo >&2
+	echo >&2 "Each line of the recipientfile looks like this:"
+	echo >&2 "\"email address\" \"survey url\" \"project\" \"revision\""
+	echo >&2
+	echo >&2 "This would be a valid line:"
+	echo >&2 "\"john.doe@example.com\" \"https://example.com/link/to/survey.php?some.params=42&other.params=42\" \"Perl\" \"v5.19.0\""
+	echo >&2
+	echo >&2 "Also, make sure to configure the script and mutt before executing it!"
 }
 
 safety_countdown () {
@@ -72,6 +73,14 @@ safety_countdown () {
 	echo
 }
 
+while getopts "hp" opt; do
+	case "${opt}" in
+		p) PRETEND="true"; shift ;;
+		h) usage; exit 0 ;;
+		'?') usage; exit 1 ;;
+	esac
+done
+
 if [ -z $1 ] || [ ! -f $1 ]; then echo "Recipient file not found!"; usage; exit 1; fi
 if [ ! -f $MAILCONFIGFILE ]; then echo "MAILCONFIGFILE not found: $MAILCONFIGFILE"; usage; exit 1; fi
 if [ ! -f $BODYTEMPLATEFILE ]; then echo "BODYTEMPLATEFILE not found: $BODYTEMPLATEFILE"; usage; exit 1; fi
@@ -81,7 +90,9 @@ if [ ! -f $BODYTEMPLATEFILE ]; then echo "BODYTEMPLATEFILE not found: $BODYTEMPL
 
 NUMTOTAL=$(wc -l $1 | awk '{ print $1 }')
 
-safety_countdown $NUMTOTAL 60
+if [ -z $PRETEND ]; then
+	safety_countdown $NUMTOTAL 60
+fi
 
 while read -r line; do
 	TIMESTAMP=$(date +%H:%M:%S)
@@ -97,8 +108,18 @@ while read -r line; do
 
 	[ $COUNT -eq 1 ] && [ $RECIPIENT == "email" ] && ( echo "$TIMESTAMP skipping header" | tee -a $LOG ) && NUMTOTAL=$[$NUMTOTAL -1] && echo && continue
 
-	# fill data into placeholders and fire
-	sed -e "s/\$PROJECT/$PROJECT/" -e "s/\$REVISION/$REVISION/" -e "s/\$URL/$URL/" < $BODYTEMPLATEFILE | tee -a ${TMPDIR}/mail.$COUNT | mutt -F $MAILCONFIGFILE -s "$SUBJECT" -b "\"$MAILFROM\"" "\"$RECIPIENT\"" 2>&1 | tee -a $MAILLOG
+	if [ $PRETEND ]; then
+		# fill data into placeholders and print message to stdout
+		echo "Subject: $SUBJECT" | tee -a ${TMPDIR}/mail.$COUNT
+		echo "From: $MAILFROM" | tee -a ${TMPDIR}/mail.$COUNT
+		echo "To: $RECIPIENT" | tee -a ${TMPDIR}/mail.$COUNT
+		echo | tee -a ${TMPDIR}/mail.$COUNT
+		sed -e "s/\$PROJECT/$PROJECT/" -e "s/\$REVISION/$REVISION/" -e "s/\$URL/$URL/" < $BODYTEMPLATEFILE | tee -a ${TMPDIR}/mail.$COUNT
+		echo
+	else
+		# fill data into placeholders and fire
+		sed -e "s/\$PROJECT/$PROJECT/" -e "s/\$REVISION/$REVISION/" -e "s/\$URL/$URL/" < $BODYTEMPLATEFILE | tee -a ${TMPDIR}/mail.$COUNT | mutt -F $MAILCONFIGFILE -s "$SUBJECT" -b "\"$MAILFROM\"" "\"$RECIPIENT\"" 2>&1 | tee -a $MAILLOG
+	fi
 	RET=$?
 	if [ $RET -eq 0 ]; then STATUS="OK"; else STATUS="ERROR"; fi
 	echo "$TIMESTAMP $STATUS $PROGRESS mail sent to "\"$RECIPIENT\"" with return code $(echo $RET)" | tee -a $LOG
